+++ Assets/PlayMaker Custom Actions/uGui/EventSystem/uGuiOnBeginDragEvent.cs
@@ -12,23 +12,23 @@ EcoMetaEnd
 
 using UnityEngine;
 using UnityEngine.EventSystems;
-
+using UnityEngine.Events;
 
 #if UNITY_5_6_OR_NEWER   
 namespace HutongGames.PlayMaker.Actions
 {
 	[ActionCategory("uGui")]
-	[Tooltip("Sends event when OnDrag is called on the GameObject. Warning this event is sent everyframe while dragging." +
-		"\n Use GetLastPointerDataInfo action to get info from the event.")]
-	public class uGuiOnDragEvent : FsmStateAction
+	[Tooltip("Sends event when GameObject is called before a drag is started." +
+		"\n Use GetLastPointerDataInfo action to get info from the event")]
+	public class uGuiOnBeginDragEvent : FsmStateAction
 	{
 		[RequiredField]
 		[Tooltip("The GameObject with the UGui button component.")]
 		public FsmOwnerDefault gameObject;
 
 		[UIHint(UIHint.Variable)]
-		[Tooltip("Event sent when OnDrag is called")]
-		public FsmEvent onDragEvent;
+		[Tooltip("Event sent when OnBeginDrag is called")]
+		public FsmEvent onBeginDragEvent;
 
 
 		GameObject _go;
@@ -38,7 +38,7 @@ namespace HutongGames.PlayMaker.Actions
 		public override void Reset()
 		{
 			gameObject = null;
-			onDragEvent = null;
+			onBeginDragEvent = null;
 		}
 		
 		public override void OnEnter()
@@ -61,8 +61,8 @@ namespace HutongGames.PlayMaker.Actions
 				entry = new EventTrigger.Entry ();
 			}
 
-			entry.eventID = EventTriggerType.Drag;
-			entry.callback.AddListener((data) => { OnDragDelegate((PointerEventData)data); });
+			entry.eventID = EventTriggerType.BeginDrag;
+			entry.callback.AddListener((data) => { OnBeginDragDelegate((PointerEventData)data); });
 
 			_trigger.triggers.Add(entry);
 		}
@@ -73,10 +73,11 @@ namespace HutongGames.PlayMaker.Actions
 			_trigger.triggers.Remove (entry);
 		}
 
-		void OnDragDelegate( PointerEventData data)
+
+		void OnBeginDragDelegate( PointerEventData data)
 		{
 			GetLastPointerDataInfo.lastPointeEventData = data;
-			Fsm.Event(onDragEvent);
+			Fsm.Event(onBeginDragEvent);
 		}
 	}
 }