+++ Assets/PlayMaker Custom Actions/uGui/EventSystem/uGuiOnDropEvent.cs
@@ -18,17 +18,17 @@ using UnityEngine.EventSystems;
 namespace HutongGames.PlayMaker.Actions
 {
 	[ActionCategory("uGui")]
-	[Tooltip("Sends event when OnDrag is called on the GameObject. Warning this event is sent everyframe while dragging." +
+	[Tooltip("Sends event when OnDrop is called on the GameObject. Warning this event is sent everyframe while dragging." +
 		"\n Use GetLastPointerDataInfo action to get info from the event.")]
-	public class uGuiOnDragEvent : FsmStateAction
+	public class uGuiOnDropEvent : FsmStateAction
 	{
 		[RequiredField]
 		[Tooltip("The GameObject with the UGui button component.")]
 		public FsmOwnerDefault gameObject;
 
 		[UIHint(UIHint.Variable)]
-		[Tooltip("Event sent when OnDrag is called")]
-		public FsmEvent onDragEvent;
+		[Tooltip("Event sent when OnDrop is called")]
+		public FsmEvent onDropEvent;
 
 
 		GameObject _go;
@@ -38,7 +38,7 @@ namespace HutongGames.PlayMaker.Actions
 		public override void Reset()
 		{
 			gameObject = null;
-			onDragEvent = null;
+			onDropEvent = null;
 		}
 		
 		public override void OnEnter()
@@ -61,8 +61,8 @@ namespace HutongGames.PlayMaker.Actions
 				entry = new EventTrigger.Entry ();
 			}
 
-			entry.eventID = EventTriggerType.Drag;
-			entry.callback.AddListener((data) => { OnDragDelegate((PointerEventData)data); });
+			entry.eventID = EventTriggerType.Drop;
+			entry.callback.AddListener((data) => { OnDropDelegate((PointerEventData)data); });
 
 			_trigger.triggers.Add(entry);
 		}
@@ -73,10 +73,10 @@ namespace HutongGames.PlayMaker.Actions
 			_trigger.triggers.Remove (entry);
 		}
 
-		void OnDragDelegate( PointerEventData data)
+		void OnDropDelegate( PointerEventData data)
 		{
 			GetLastPointerDataInfo.lastPointeEventData = data;
-			Fsm.Event(onDragEvent);
+			Fsm.Event(onDropEvent);
 		}
 	}
 }