+++ Assets/PlayMaker Custom Actions/uGui/EventSystem/uGuiOnSelectEvent.cs
@@ -1,34 +1,25 @@
 ﻿// (c) Copyright HutongGames, LLC 2010-2017. All rights reserved.
 // __ECO__ __PLAYMAKER__ __ACTION__ 
-/*
-EcoMetaStart
-{
-	"script dependancies":[
-		"Assets/PlayMaker Custom Actions/uGui/EventSystem/GetLastPointerDataInfo.cs"
-	]
-}
-EcoMetaEnd
-*/
 
 using UnityEngine;
 using UnityEngine.EventSystems;
-
+using UnityEngine.Events;
 
 #if UNITY_5_6_OR_NEWER   
 namespace HutongGames.PlayMaker.Actions
 {
 	[ActionCategory("uGui")]
-	[Tooltip("Sends event when OnDrag is called on the GameObject. Warning this event is sent everyframe while dragging." +
-		"\n Use GetLastPointerDataInfo action to get info from the event.")]
-	public class uGuiOnDragEvent : FsmStateAction
+	[Tooltip("Sends event when Called by the EventSystem when a Select event occurs." +
+		"\n Use GetLastPointerDataInfo action to get info from the event")]
+	public class uGuiOnSelectEvent : FsmStateAction
 	{
 		[RequiredField]
 		[Tooltip("The GameObject with the UGui button component.")]
 		public FsmOwnerDefault gameObject;
 
 		[UIHint(UIHint.Variable)]
-		[Tooltip("Event sent when OnDrag is called")]
-		public FsmEvent onDragEvent;
+		[Tooltip("Event sent when OnSelect is called")]
+		public FsmEvent onSelectEvent;
 
 
 		GameObject _go;
@@ -38,7 +29,7 @@ namespace HutongGames.PlayMaker.Actions
 		public override void Reset()
 		{
 			gameObject = null;
-			onDragEvent = null;
+			onSelectEvent = null;
 		}
 		
 		public override void OnEnter()
@@ -55,14 +46,14 @@ namespace HutongGames.PlayMaker.Actions
 			{
 				_trigger = _go.AddComponent<EventTrigger>();
 			}
-
+				
 			if (entry == null)
 			{
 				entry = new EventTrigger.Entry ();
 			}
 
-			entry.eventID = EventTriggerType.Drag;
-			entry.callback.AddListener((data) => { OnDragDelegate((PointerEventData)data); });
+			entry.eventID = EventTriggerType.Select;
+			entry.callback.AddListener((data) => { OnSelectDelegate((PointerEventData)data); });
 
 			_trigger.triggers.Add(entry);
 		}
@@ -73,10 +64,11 @@ namespace HutongGames.PlayMaker.Actions
 			_trigger.triggers.Remove (entry);
 		}
 
-		void OnDragDelegate( PointerEventData data)
+
+		void OnSelectDelegate( PointerEventData data)
 		{
 			GetLastPointerDataInfo.lastPointeEventData = data;
-			Fsm.Event(onDragEvent);
+			Fsm.Event(onSelectEvent);
 		}
 	}
 }